snapshot

+++ Ecommerce.Entity/Views/CustomerStats.cs
@@ -5,10 +5,10 @@ public class CustomerStats
     public uint CustomerId { get; set; }
     public decimal? TotalSpent { get; set; }
     public decimal? TotalDiscountUsed { get; set; }
-    public int? TotalOrders { get; set; }
-    public int? TotalReviews { get; set; }
-    public int? TotalKarma { get; set; }
-    public int? ReviewVotes { get; set; }
-    public int? CommentVotes { get; set; }
-    public int? TotalComments { get; set; }
+    public uint? TotalOrders { get; set; }
+    public uint? TotalReviews { get; set; }
+    public long? TotalKarma { get; set; }
+    public uint? ReviewVotes { get; set; }
+    public uint? CommentVotes { get; set; }
+    public uint? TotalComments { get; set; }
 }
